requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
httpcore==1.0.5
qdrant-client==1.9.1
grpcio==1.75.0
grpcio-tools==1.75.0
pydantic==2.11.9
portalocker==2.10.1
beautifulsoup4==4.12.3
lxml==5.3.0
loguru==0.7.2
win32-setctime==1.2.0
prometheus-client==0.21.0
FlagEmbedding==1.3.3
numpy==1.26.4
tqdm==4.66.5
rapidfuzz==3.9.7
uvicorn==0.30.6
gunicorn==22.0.0
fastapi==0.115.6
starlette==0.41.2
uvicorn[standard]==0.30.6
playwright==1.55.0
transformers==4.44.2
sentence-transformers[onnx-gpu]==3.0.1
onnx
onnxruntime-directml
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
torch-directml
telegramify-markdown==0.1.0
pytelegrambotapi==4.29.1
mistletoe==1.4.0
marshmallow==3.23.1
redis==5.0.8
